+++ src/main/java/io/github/gerritsmith/financeapp/dto/form/DeliveryLegFormDTO.java
@@ -1,4 +1,4 @@
-package io.github.gerritsmith.financeapp.dto;
+package io.github.gerritsmith.financeapp.dto.form;
 
 import io.github.gerritsmith.financeapp.model.DeliveryLeg;
