+++ src/main/java/io/github/gerritsmith/financeapp/dto/form/ExpenseFormDTO.java
@@ -1,4 +1,4 @@
-package io.github.gerritsmith.financeapp.dto;
+package io.github.gerritsmith.financeapp.dto.form;
 
 import io.github.gerritsmith.financeapp.model.Expense;
 import org.springframework.format.annotation.DateTimeFormat;
