+++ src/main/java/io/github/gerritsmith/financeapp/dto/form/ShiftFormDTO.java
@@ -1,4 +1,4 @@
-package io.github.gerritsmith.financeapp.dto;
+package io.github.gerritsmith.financeapp.dto.form;
 
 import io.github.gerritsmith.financeapp.model.Shift;
 import org.springframework.format.annotation.DateTimeFormat;
